+++ R/class-precision.R
@@ -189,7 +189,7 @@ warn_precision_undefined_binary <- function(event, count) {
     message = message,
     events = event,
     counts = count,
-    .subclass = "yardstick_warning_precision_undefined_binary"
+    class = "yardstick_warning_precision_undefined_binary"
   )
 }
 
@@ -209,14 +209,14 @@ warn_precision_undefined_multiclass <- function(events, counts) {
     message = message,
     events = events,
     counts = counts,
-    .subclass = "yardstick_warning_precision_undefined_multiclass"
+    class = "yardstick_warning_precision_undefined_multiclass"
   )
 }
 
-warn_precision_undefined <- function(message, events, counts, ..., .subclass = character()) {
+warn_precision_undefined <- function(message, events, counts, ..., class = character()) {
   rlang::warn(
     message = message,
-    .subclass = c(.subclass, "yardstick_warning_precision_undefined"),
+    class = c(class, "yardstick_warning_precision_undefined"),
     events = events,
     counts = counts,
     ...