+++ R/class-sens.R
@@ -228,7 +228,7 @@ warn_sens_undefined_binary <- function(event, count) {
     message = message,
     events = event,
     counts = count,
-    .subclass = "yardstick_warning_sens_undefined_binary"
+    class = "yardstick_warning_sens_undefined_binary"
   )
 }
 
@@ -248,14 +248,14 @@ warn_sens_undefined_multiclass <- function(events, counts) {
     message = message,
     events = events,
     counts = counts,
-    .subclass = "yardstick_warning_sens_undefined_multiclass"
+    class = "yardstick_warning_sens_undefined_multiclass"
   )
 }
 
-warn_sens_undefined <- function(message, events, counts, ..., .subclass = character()) {
+warn_sens_undefined <- function(message, events, counts, ..., class = character()) {
   rlang::warn(
     message = message,
-    .subclass = c(.subclass, "yardstick_warning_sens_undefined"),
+    class = c(class, "yardstick_warning_sens_undefined"),
     events = events,
     counts = counts,
     ...