+++ tests/testthat/helper-data.R
@@ -108,3 +108,18 @@ data_three_class <- function() {
 data_soybean <- function() {
   readRDS(test_path("data/helper-soybean.rds"))
 }
+
+# Made up data
+data_counts <- function() {
+  count_truth <- c(2L,   7L,   1L,   1L,   0L,  3L)
+  count_pred  <- c(2.14, 5.35, 1.65, 1.56, 1.3, 2.71)
+  count_results <- dplyr::tibble(count = count_truth, pred = count_pred)
+
+  count_poor <- count_results
+  count_poor$pred <- 0.1
+
+  count_missing <- count_results
+  count_missing$pred[1] <- NA
+
+  list(basic = count_results, missing = count_missing, poor = count_poor)
+}