+++ NEWS.md
@@ -1,3 +1,5 @@
+# yardstick (development version)
+
 # yardstick 0.0.8
 
 * New `metric_tweak()` for adjusting the default values of optional arguments in